partially updated joystick stuff, still have to move smm to somatic
git-svn-id: file:///home/svn/humanoids/drivers/joystick@3920 4f504ea5-f356-0410-adec-f1b3679f3bb7
