+++ snachd.c
@@ -60,11 +60,9 @@
 
 
 /* Option Vars */
-static int opt_sndev = 0;
 static int opt_verbosity = 0;
 static int opt_create = 0;
-static char *opt_ach_chan = SPACENAV_CHANNEL_NAME;
-static int opt_axis_cnt = SNACH_NAXES;
+static const char *opt_ach_chan = SPACENAV_CHANNEL_NAME;
 
 /* ---------- */
 /* ARGP Junk  */
@@ -135,6 +133,35 @@ static int parse_opt( int key, char *arg, struct argp_state *state) {
 /* Function Defs */
 /* ------------- */
 
+/**
+ * Block, waiting for a mouse event
+ */
+void snach_read_to_msg(Somatic__Joystick *msg, spnav_event *spnevent)
+{
+	spnav_wait_event(spnevent);
+
+	if (spnevent->type == SPNAV_EVENT_MOTION) {
+			msg->axes->data[0] = (double)spnevent->motion.x /  SPNAV_MOTION_MAX;
+			msg->axes->data[1] = (double)spnevent->motion.y /  SPNAV_MOTION_MAX;
+			msg->axes->data[2] = (double)spnevent->motion.z /  SPNAV_MOTION_MAX;
+			msg->axes->data[3] = (double)spnevent->motion.rx / SPNAV_MOTION_MAX;
+			msg->axes->data[4] = (double)spnevent->motion.ry / SPNAV_MOTION_MAX;
+			msg->axes->data[5] = (double)spnevent->motion.rz / SPNAV_MOTION_MAX;
+	}
+	else if (spnevent->type == SPNAV_EVENT_BUTTON) {
+		if (spnevent->button.bnum == 0) {
+			msg->buttons->data[0] = (int64_t)spnevent->button.press;
+		}
+		else if (spnevent->button.bnum == 1)  {
+			msg->buttons->data[1] = (int64_t)spnevent->button.press;
+		}
+		else {
+			//TODO remove this once it's tested:
+			printf("Shouldn't be here\n");
+			exit(-1);
+		}
+	}
+}
 
 /* ---- */
 /* MAIN */
